+++ packages/babel-plugin-rn-stylename-to-style/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0-alpha.7](https://github.com/startupjs/startupjs/compare/v0.3.0-alpha.6...v0.3.0-alpha.7) (2026-04-16)
+
+**Note:** Version bump only for package @cssxjs/babel-plugin-rn-stylename-to-style
+
+
+
+
+
+# [0.3.0-alpha.6](https://github.com/startupjs/startupjs/compare/v0.3.0-alpha.5...v0.3.0-alpha.6) (2026-04-12)
+
+**Note:** Version bump only for package @cssxjs/babel-plugin-rn-stylename-to-style
+
+
+
+
+
+# [0.3.0-alpha.5](https://github.com/startupjs/startupjs/compare/v0.3.0-alpha.4...v0.3.0-alpha.5) (2026-04-09)
+
+**Note:** Version bump only for package @cssxjs/babel-plugin-rn-stylename-to-style
+
+
+
+
+
+# [0.3.0-alpha.4](https://github.com/startupjs/startupjs/compare/v0.3.0-alpha.3...v0.3.0-alpha.4) (2026-04-06)
+
+**Note:** Version bump only for package @cssxjs/babel-plugin-rn-stylename-to-style
+
+
+
+
+
+# [0.3.0-alpha.1](https://github.com/startupjs/startupjs/compare/v0.3.0-alpha.0...v0.3.0-alpha.1) (2026-03-25)
+
+
+### Performance Improvements
+
+* update babel pug plugin to short-circuit processing of files which don't have pug ([6317ee3](https://github.com/startupjs/startupjs/commit/6317ee3d9b30a218cc6a5157327ad85fa46e1ac2))
+
+
+
+
+
+# [0.3.0-alpha.0](https://github.com/startupjs/startupjs/compare/v0.2.33...v0.3.0-alpha.0) (2026-03-25)
+
+
+### Features
+
+* use the new pug babel plugn and eslint plugin from 'react-pug' with support for 'style' tags and full TSX support ([28e7a22](https://github.com/startupjs/startupjs/commit/28e7a22e8069dec77c4a120f42b951ef90a4bef9))
+
+
+
+
+
 ## [0.2.32](https://github.com/startupjs/startupjs/compare/v0.2.31...v0.2.32) (2026-01-25)
 
 **Note:** Version bump only for package @cssxjs/babel-plugin-rn-stylename-to-style